Indexed dataset + refactoring
